requirements.txt
django
djangorestframework
django-extensions
ruff>=0.1.0
pytest>=7.0.0
pytest-django>=4.0.0
djangorestframework-simplejwt
drf-spectacular
django-filter
python-dotenv
